+++ CHANGELOG.md
@@ -4,11 +4,18 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.92.0] 2022-07-29
+
+- New command hardis:org:retrieve:source:analytics to retrieve all analytics (CRM Analytics/TCRM) sources
+- New deployment tips (Wave analytics)
+- Fix writePackageXml method when there is not an existing file
+
 ## [2.91.0] 2022-07-15
 
 - Fix issue when force:source command wrappers arguments contain spaces [(#269)](https://github.com/hardisgroupcom/sfdx-hardis/issues/269))
 - Upgrade [MegaLinter](https://oxsecurity.github.io/megalinter/latest/) to v6
 - Upgrade yarn dependencies
+
 
 ## [2.90.0] 2022-06-24